# Add inline to `Mul<PathEl>` implementation for `Affine`